requirements.txt
openai==1.8.0
langchain==0.1.1
langchain-community==0.0.13
langchain-openai==0.0.2.post1
streamlit==1.30.0
streamlit-chat==0.1.1
pypdf==3.17.4
chromadb==0.3.29
tiktoken==0.5.2
python-dotenv==1.0.0
holidays==0.41